Fix T46553: FCurve Modifier parameter is changed only with time change
UI buttons are missing update callback, which is copied from the RNA
update callback.

The delete modifier callback is also missing an update to recompute the
parent ID (object most likely) data, though it needs a way to only
update the parent ID (for now the whole scene is updated).

Reviewers: #animation

Differential Revision: https://developer.blender.org/D2118